Split Geometry2D into its core API and a helper SegmentGrid that holds line segments inside a Grid. 3D will use segments for both walls and seams (where regions connect). Add unit tests for SegmentGrid.
